Added codegen for a new app.
This commit adds codegen so you can run `make app` and have it generate
all of the necessary components. It does require golang + make to be
able to work though I suspect that's not a huge lift. If it becomes one,
we can add it to pixlet and release the creation as part of that binary.