Fix big clock summary.
This commit shortens big clocks summary. The build didn't run on the PR
(my fault, the action was misconfigured) that would have caught this.
The reason it has a max length is a problem is it shows up directly in
the UI. We may be able to support longer then 27 characters, but we'd
need to test it on multiple devices to see how it looked.