build: Add linter support (#63)
* Added starlark linter to the build.

This commit adds a starlark linter to ensure the code is properly
formatted so we can land changes with confidence.

* Formatted apps.

This commit formats the apps directory using:
```
buildifier -r ./
```

* Fix they said so.

This commit fixes a weird buildifier change that didn't handle a
character properly. I spot checked the rest of the changes to make sure
there are no other weird changes.