bugfix: Fix moonphase app.
This commit fixes an issue where moonPhase or phaseImage might not be
defined by setting them to the first phase before attempting to find the
correct phase. There is likely a better way to do this but I wanted to
do the minimal changeset to get the tests to pass.