Add 'HVV Departures' app.
This app displays real-time departure times for trains, buses and ferries in Hamburg (HVV).

It has a lot of configuration options including:

* Picking a station via typeahead.
* Picking a direction via typeahead.
* Picking a departure time formats:
  - Relative (e.g. "3 min")
  - Absolute 24-hour format + delay indicator (e.g. "09:30+3")
  - Absolute 12-hour format + delay indicator (e.g. "01:30 PM+2")
* Picking a departure time offset, if it takes some time to get to the departure station by foot.
  - now
  - in 5 minutes
  - in 10 minutes
  - in 15 minutes
* Selecting which means of transport to include:
  - Subway (U-Bahn)
  - Suburban trains (S-Bahn)
  - Commuter trains (AKN)
  - Buses (MetroBus, XpressBus, NachtBus)
  - Regional (express) trains (Regional-Bahn, Regional-Express)
  - Ferries

For all of the above means of transport it knows how to render the respective icons with
the correct shapes and colors, to make it easier to see which line is departing at a glance.