coingecoprice: Hardcode supported coins and currencies
We were generating too many requests against the API when running tests,
to the point where we'd get blocked. Instead, just hardcode this data.

It will have to be periodically updated, and/or we need to expose
something in Starlark to tell apps that they are running in a unit test
so they can avoid scenarios like this.