Sports Standings: Bug fixes (#302)
Noticed the other day the app wasn't rendering. Edits here should fix them.
Should have known indexing by location and not name was a bad idea. With seasons starting and ending, the json files are a smidge different.

Fully expect to have the same issue during the offseason for NHL and NBA, but will check that when the time comes. I also have back up sources for MLB and NHL as far as standings are concerned in case this breaks again.

Sorry for all the fixes lately!