On two week chart, change all of last weeks bars to dark grey
This was kind of cluttered before since, for last week, also it showed a diferent colored bar based on whether the goal was hit or not.  I decided to clean up the display to make this weeks data more prominent. Last week's data is consistently just displayed as a dark grey bar, regardless of whther or not the XP goal was achieved.