Fix and improve the cache_key.
As a followup to the previous change, when fetching global earthquakes, the cache_key can be set to just the magnitude since the other parameters are irrelevant. This will allow for more cache hits across the community. While I was here, I slightly adjusted the format string to make the cache_key a little more readable during debugging.