New app: USGS Earthquakes (#331)
* Add USGS Earthquakes app

* cleanups

* small fix

* Added three additional radius options.

This change adds three additional options for the radius setting:

  2500km
  5000km
  Unlimited

The last option allows the user to show earthquakes worldwide. For this last option,
it probably makes sense to use it with a larger magnitude, such as 3 or higher.

Credit to rjison@ for suggesting these additions.

* Fix and improve the cache_key.

As a followup to the previous change, when fetching global earthquakes, the cache_key can be set to just the magnitude since the other parameters are irrelevant. This will allow for more cache hits across the community. While I was here, I slightly adjusted the format string to make the cache_key a little more readable during debugging.