mbta: Fix mbta stop options.
This commit resolves an issue where the stop is returned as an option.
There are two issues on the Tidbyt side that we need to address. One is
that we need to support location options in the dev tools. The second is
it's really confusing that one returns a schema.Option and gets back a
JSON string.