Update defaults for better previews.
While the default options worked well for California, they probably aren't very good for other parts of the world. Changed the default radius to unlimited which should always result in a good preview. I likely could have set the minimum magnitude to 4, but we'll use 3 to be extra sure we'll get at least one result.