Remove apostrophe from summary
Hmmm, don't love this, but it looks like things fail when the apostrophe is present.

Error is "app summaries should start with an uppercased character", which is weird because it did.

The same thing originally happened when I ran make app. It only succeeded when I dropped the apostrophe. I went back and put it in manually. Out it comes again.