New app: Tartan
Weaves and displays a random tartan

Probably a bit cleverer than it needs to be, as it weaves pixels in detail, and then loses all the detail when resizing to fit on a tidbyt screen.

Has a library over 3,000 patterns to choose from. Default mode is to display a random one each time it runs.

Nice way for Tidbyt users to connect with their Scottish heritage, or just enjoy some pleasing low-resolution patterns.