Add Elizabeth Line
It opened last week so it wasn't possible to try out before.

Transport for London have recently said they don't treat it as an underground line, and so it's classed as a separate mode of transport in the API.

https://www.standard.co.uk/news/london/crossrail-elizabeth-line-tube-tfl-london-underground-b1001058.html

And TFL Rail is no longer returns data, as it was rebranded as part of the new Elizabeth Line.