New app: Tartan (#459)
* New app: Tartan

Weaves and displays a random tartan

Probably a bit cleverer than it needs to be, as it weaves pixels in detail, and then loses all the detail when resizing to fit on a tidbyt screen.

Has a library over 3,000 patterns to choose from. Default mode is to display a random one each time it runs.

Nice way for Tidbyt users to connect with their Scottish heritage, or just enjoy some pleasing low-resolution patterns.

* Fixes for failing inputs

Ran a script to generate images for all the patterns, and found:

1. Some were missing colour definitions. Added obvious ones like "w" for white and "k" for black. Removed one pattern where I had no idea.

2. Avoid divison by zero if pattern is smaller than the screen (particularly looking at Moncreiffe)

3. Fill the screen properly if pattern is smaller. Before, I was doing a single repeat but Moncreiffe needs multiple.

4. Fix regex to handle patterns that don't contain a pivoted part (i.e. there are no middle brackets).

60MB archive containing all the patterns available on dropbox. Long text will scroll on tidbyt, but not in these stills.
https://www.dropbox.com/s/e6fuk7oocmzzd70/tiny_tartans.zip?dl=0

* Allow hiding the name

Maybe you just want a nice, anonymous pattern on a rainy evening.