Pokedex: Swap Random function
Swapped the random function with the new `random.number` module available in pixlet. Not sure if this will help with the app constantly repeating a lot of the starter pokemon in the evening (at least in my experience).

However this should ideally be tested before actually making it live. I also only commented out the original random module stuff just in case that wasn't the issue. FWIW locally it seems to randomize, but doesn't always on the tidbyt it seems.