New App: Shuffle Images (#510)
Randomly display an image from a list of up to 20 user-defined images.

Still trying to figure out how to get them to display linearly (i.e. order defined), but can't figure out a way to cache an index that wouldn't get constantly overwritten.
A unique index key would be fine across devices since the images aren't cached, but couldn't figure out a good way to do that at the moment.

This app also may help reduce the number of apps listed on the device if you have a lot of photos.