apps: Update icons.
This commit updates all of the icons so the build passes. In many
instances, these were icons taken from examples that are no longer valid
and I simply updated them to a drop in replacement. The rest of the
instances were icon names that simply were not valid and I did my best
to find an appropriate replacement.