feat: Icon validation (#565)
* apps: Add icons test.

This commit adds a map of all of the font awesome icons we support. This
map is the same map that will be released inside of our mobile app, so
the added test will ensure all apps in the community repo have icons
that work.

* apps: Update icons.

This commit updates all of the icons so the build passes. In many
instances, these were icons taken from examples that are no longer valid
and I simply updated them to a drop in replacement. The rest of the
instances were icon names that simply were not valid and I did my best
to find an appropriate replacement.

* Fix linter