Add Shabbat times app
There was a request on the discord for a simple app that shows the start and end times of Shabbat, the Jewish day of rest.
https://discord.com/channels/928484660785336380/1011854621486092298

This is very plain, which also seems appropriate. Just a blob of text.

What it shows:
 - from early Sunday to start on Friday: time it will start
 - from start time to end on Saturday: time it will end
 - between end time on Saturday and midnight: time it ended

Corroborated times with what I get in my location from Google's search results.

There's different methods for calculating the end time. I've raised in PR for the go-sunrise library that pixlet depends on, which would allow another option. If I don't hear anything, I'll just add that to pixlet's sunrise module.
https://github.com/nathan-osman/go-sunrise/pull/9