pulsechain: Error handling and new API endpoint (#623)
The previous version did not handle when the remote API server returned unexpected data.

data is always returned if the API call is successful. This commit checks for that property. If it is not returned, then instead of throwing an error, a placeholder indicating no data will be used instead.

Also, I updated the remote API to the only uniswap v3 subgraph that is working.