Show destinations
Shows where each bus is going, because that's often quite helpful, especially if the bus isn't going the whole way to the end of the line.

Limited in how I can animate this when the app already includes a marquee. It's set to toggle between texts after 100 frames, which should be ~2.5 seconds

We can also avoid the header being so long and having to scroll so much now. It doesn't have to state a direction for the stop, since we have directions for each individual bus.