Improve London transport apps (#671)
* Show line on Tube app

Relying on colour alone isn't great, as I found out when I showed it to a colourblind friend.

Long station names now scroll horizontally rather than taking up the whole box.

* Add sequential rendering mode

It's quite distracting to have a long, colourful block of text constantly scrolling.

Added a new default display mode that shows two lines at a time, and only swaps every 2 seconds, so it all still fits into the 15 second loop.

* Show destinations

Shows where each bus is going, because that's often quite helpful, especially if the bus isn't going the whole way to the end of the line.

Limited in how I can animate this when the app already includes a marquee. It's set to toggle between texts after 100 frames, which should be ~2.5 seconds

We can also avoid the header being so long and having to scroll so much now. It doesn't have to state a direction for the stop, since we have directions for each individual bus.