Encrypt using correct file name
Not sure if I had done this right first time, but now the app keys are definitely encrypted using "london_bus_stop" and "tube_status" rather than "londonbusstop" and "tubestatus".

Probably hard to tell if I had it wrong before because the API seems to allow a certain amount of freebie queries without an API key on these endpoints.