Avoid spamming on API failures
We cache good data so that we don't overload the API or exceed our quota. But when the response is bad, we'll retry every time. That's likely to happen when something's gone wrong (like yesterday when crowdedness data wasn't available). We don't want to make a bad situation worse for them.