Handle API failures better (#686)
* Handle errors better in tube

1. Don't search for locations outside the UK because the API won't accept them

2. Don't fail. Instead show "no data" screen

* Handle errors better in tube_status

Show error screen instead of fail() and skip unexpected data instead of failing()

* Handle errors better in london_bus_stop

1. Don't request locations outside UK bounds since the API returns an failure

2. Remove fail() and show message instead

* Handle errors better in busytube

1. Show "unknown" rather than fail() in case of missing data

2. Don't ask for locations outside the UK as the API returns a failure

3. Fix option name as the line doesn't matter here

* Handle errors better in borisbikes

1. Don't ask for locations outside the UK as the API won't give a response

2. Show result saying "No data" rather than fail()

* Avoid spamming on API failures

We cache good data so that we don't overload the API or exceed our quota. But when the response is bad, we'll retry every time. That's likely to happen when something's gone wrong (like yesterday when crowdedness data wasn't available). We don't want to make a bad situation worse for them.

* Fix line text colour

Noticed this with Farringdon - Circle and Farringdon - Hammersmith & City. The station name was black as intended. It was harder to read as well as inconsistent.