New app: National Rail (#708)
* New app: National Rail

Live departure boards for UK mainline railway services. Two display modes! Filter by destination! Live data from National Rail Enquiries!

* Pass test when can't decrypt

Add a bit of indirection around the use of encrypted secrets so that the tests pass when secret.decrypt returns None.

* Lint :(

* Halve API queries

Don't need to make separate queries for the two display modes. The only difference is the number of rows fetched. But one is a subset of the other, so fetch and cache the longer one.

* Strip full stop in 16 character name

Some have it and it looks rubbish. Also fix default option because it led to unexpected failures in typical case where station list is available.