Better buses (#704)
* More compact London buses

Using smaller font similar to https://github.com/tidbyt/community/pull/703, because I found it wasn't fitting enough of the destination name on screen. This also lets it show a fourth bus, which is nice.

* Show more buses

Worked out how to animate between multiple screens so we can now show up to 12 buses rather than just two.

* Don't cut off destination text

WrappedText won't include the next word if it's too long. But we may as well show it even if it's cut off, as it can help disambiguate.