Sort station list by location
glovepup on discord is right. It is a pain to scroll through >2500 stations to pick the one you want.

I downloaded the 35MB station list XML file and parsed out the key bits of info.

With latitude and longitude you can search for an address and get stations near it, rather than alphabetical.