Top Crypto Prices
This app shows the latest prices (USD) and changes for the top 10 cryptocurrencies.

It is a very simple app, but does exactly what I needed for myself, and why I bought the Tidbyt in the first place. Since there was no app that could do what I had hoped for, I wrote one myself.