deps: Bump pixlet version. (#855)
* deps: Bump pixlet version.

For some reason, the rennovate build for this PR was failing.

* ci: Update build to use go 1.19.

The go.mod already uses go 1.19, this commit updates our build to use
the same version.

* go: Fix the linter.

The linter was complaining because io/util functionality has been moved
into the os package. This commit uses the os package instead.

* deps: Update everything.

This commit runs go get -u ./... to update the rest of our outstanding
dependencies.