Change NS Error Cache to 30 mins (#859)
This cache is used only when Nightscout returns an error. The cache ensures a blood sugar value is shown (and displayed as state) for up to 30 minutes if NS is temporarily unavailable. It does not affect live glucose values if they are available as this app never displays cached data unless there is a NS error.