New App: GitHub Repo
Display stats for a GitHub repository.

the auth key is just done by creating a personal access token (classic) and the dev key is set to never expire currently.

One big improvment that is necessary is the consistency of the icons and color, but I'm trying to figure out sourcing these with consistency.