Fix NES Quotes unused variable.
The original method was not correct, and the linter complained about
result being unused. The issue is, this method is used to create an ID
for schema fields. If we change the ID, it will break any existing
installations. Therefore, the safest option is to mimic the behavior of
the original function.