Fix ethstaker lint warning.
This commit resolves our last lint waring for status being undefined. I
am not sure the intent here, and looking through the rest of the source
code was no help. I presume this method already was erroring given this
was not defined. My fix satisfies the linter but it likely does not
resolve the logic error in this mehthod.