fix: Set max_age for clocks. (#981)
By default, the Tidbyt will display stale images it recieves near idefinitely if it cannot get an update from our backend. In almost all cases, this is the desired behavior. The weather isn't going to change much within a 10 minute time frame. However, a 10 minute stale clock would be incredibly frustrating. We implemented a mitigation for our clock some time ago that sets the max age to two minutes for the clock, ensuring the most stale clock image the device would show would be no greater then 2 minutes. However, we never rolled out that change to all clocks. This commit updates all clocks in the community repo to not be displayed if they are stale.