Fido: Allow the ability to randomize the actions Fido takes
I think there is a bug in either `pixlet serve` or something because I had to set random=="true" instead of it being a boolean. So am having to check against the string. Also required a toggle of True/False to get the extra config to show up when using `pixlet serve`