New Apps: Fixed Cache key
Accidentally was setting the cache key of new apps differently than the key I was trying to get, so it was constantly hitting the tidbyt api. This app will not display anything until new apps are merged. It seemed to be working yesterday, so we'll see if this works. I think the logic on line 37 should be right (we want it to update a current list if it just got a new old list).