build: Add git reference printing. (#1077)
I'm working to add target determination so we only build what's changed
on pull requests. In order to do that, I need to know what references
are available on PRs. It's especially confusing because I need to know
what the build system sees on forks. This commit adds a reference
printing so I can see what PRs have available today.