build: Add target determination output (#1084)
* build: Fix branch name matching.

This commit resolves an issue where a PR to a PR wouldn't trigger the
build due to there being a forward slash in the target branch name.

* build: Add target determination output.

This commit adds target determination output. The checkout action from
github uses a merge commit. This change ensures we get the parents from
the merge commit to use as the old/new inputs to pixlet.