currencyconverter: Display Modifications  (#1075)
Adjust the number of decimals displayed based on the the given number. For exchange rates where one currency is for 1000X the other, we need many decimals to show appropriate significant digits of the less valuable currency, and only 2 for the more valuable currency, as any more is silly.

What I've tried to do is have the decimals line up for most currencies and provide an appropriate number of decimals. However, for some valuations, this didn't work so I made an algorithm to provide the best display given the exchange rate.

Also, to avoid having to use # buildifier: disable=integer-division above my 1/x calculation where it wants to change my code to round off the numbers, instead I've used math.pow(x, -1) to achieve the same result, but without the error message.