build: Use merge base to figure out old commit.
This commit goes back to the merge commit as it doesn't make much
difference. What does make a difference is what two commits we diff
against. The base ref we get from GitHub is the current HEAD of main,
not what we branched off from. So what we really want is the merge base
to find the common ancestor and build what changed on top of that.