detect stale cache and invalidate after 3 cycles
if a buoy goes down it will return a 404.  Instead of displaying an error, used old data and keep track of long we are using it. Once its 30 minutes stale(or 3 data fetch cycles) invalidate and display the error.
Change error to indicate its a buoy.
Change default buoy to one that is not down at the moment.