Fix nationalrail app (#1225)
Responding to bug pointed out by users here, sorry it's taken a while:
https://discuss.tidbyt.com/t/national-rail/4485/2

The thread was helpful for pointing out that the issue occurs only when setting a filter. I found I got an error whenever a filter station was selected.

That's because I was putting a whole JSON object into the filter condition, rather than just the relevant CRS value. Tested this locally and no longer get an error when setting a filter.