New app: planetarium
Animates the positions of the planet around the sun between 1900 and 2100.

Angles around the sun should be roughly right, using data provided by NASA. The biggest source of error is from rounding to fit onto the little screen.

Relative distances from the sun to planets are way off, to make them all fit on the screen.

Script is huge, but it's mostly just lots of data. Please feel free to say if you'd like me to cut it down.