New App: Cost Calculator
Per a request on the discourse channel. This app shows the elapsed time and calculates a cost given a specified rate. It only shows hours and minutes. Would be great if there was a toggle on the datetime to only have to choose the time and not date and time.