Bug fix: ATP Tennis & WTA Tennis (#1505)
* Bug fix: ATP Tennis

Fixed bug which appears when player who is serving is not being provided by data feed. Code now checks if that data is present before showing it, or not

* Bug fix: WTA Tennis

Fixed bug which appears when player who is serving is not being provided by data feed. Code now checks if that data is present before showing it, or not