Update MVV app: Fix occasional missing departures. (#1919)
Because of the cap of max 40 departures returned by the MVV service,
we could be missing departures we were interested in if filtering down
the result set ended up skipping beyond all those 40. We now first
query the available lines for the target stop, filter them down to
what's interesting, and then retrieve departures only for the lines
remaining. The cap of 40 is no longer a problem then.

This change also includes a bit of cleanup and refactor.