Unfortunate Update
Despite my best attempts, many features were removed for the sake of function. The removed features include:

>The selection of different classes within the selected sport
>The selection of time within the selected sport
>Certain classes within sports for the sake of render time

A new feature was added: the selection of a game number

Summary:

I was unable to properly implement the intended features. The app now functions as intended, but with a few drawbacks. Instead of cycling through different games, users must select the game they with to display. Games can no longer be selected by class and a few have been removed.