New app: Thanksgiving Countdown
Adds a new app called Thanksgiving Day that is a simple daily countdown until the Thanksgiving holiday in the U.S. It includes a simple turkey GIF animation that we created.

This is my first Tidbyt app submission so happy to fix any errors or issues. Having a great time experimenting with the platform and love our Tidbyt!