Bugfix for mobygames app - corrected app-id for encrypted secret
Bugfix for mobygames app - corrected app-id for encrypted secret
I had neglected to use the correct app id when encrypting the secret,
using the id `mobygames` instead of `moby-games` as it is specified
in the manifest.yml. I've now corrected this issue.

I also modified the api key logic in the main function in order
to better produce useful debugging log messages, though I couldn't
use this to verify decryption behavior while running it locally
(since the decrypt behavior doesn't work locally).

Finally, I also improved error handling so that if we failed to
decrypt the key, we'd avoid calling MobyGames entirely
instead of calling and receiving a 401.