New app: Cuneiform (#2190)
Similar to https://github.com/tidbyt/community/pull/2179 and https://github.com/tidbyt/community/pull/2185 but for another ancient script.

Shows a new cuneiform sign every 15 seconds, plus it's name (according to Unicode) and transliteration in Sumerian. I wish I had the sources or knowledge to also support the Akkadian language, but I leave that as an exercise for the avid user.

There is some text mangling to make it display on Tidbyts. Notably, subscripted numbers aren't subscripts and `ḫ` is displayed as `ħ`. But I'm impressed as I expected many fewer of the necessary characters to work in any of pixlet's fonts.