Fix infinite height (#2499)
If the government count is exactly divisible by 13 then there isn't a remainder. So we shouldn't add a remainder column to the picture. But currently, we are trying to add a very tall column (probably due to division by almost zero), which looks weird and means the seat counts don't get shown