Add FRC ranking app to community apps (#2953)
* Add FRC ranking app to tidbyt community

This FRC ranking app will display the current ranking position of a team in the FRC competition at their active competion. The app will display the team avatar (if provided to FIRST), team number, team name, and ranking position.

* Remove encoding module from app

It is not actually used anywhere, so might as well not load it

* Add encrypted read TBA token

This should make it optional for the user to provide their own TBA token to the app

* 🚨: Apply linting and formating suggestions

* 💄: Add a default/demo view if there is no team number or tba api key passed to app

* ⚡️: QOL improvements during a comp